Schlib: use polylines instead of rectangles
The rectangle graphics primitive is a legacy item. Newly saved symbols
encode them as polylines.
Add code to convert legacy rectangles into polylines.
Add functions to check if a polyline is a rectangle.
Add function to find the center rectangle of a symbol.
Update rules S3.1 and S3.3 with the above functions.