Add a test for the kicad_sym write functionality
The basic idea is to load any kicad_sym file, convert it into the python
structure and then write it out again.
Then normalize that file and the source file so they can be compared.

Currently this still failes because kicad does not store the units sorted
by unit number.