Fix the way symbols are parsed
The name of the symbol is no longer prefixed with the library name, we need to
adjsut our parser for that. The regexp that searched for ':' is no longer
needed.
Also, the parsing of subsymbols (units) was done in a stupid way.
Now explicity only search for symbols in the top-level of the s-expr. When we
symbol is found, look for its units the same way we look for its properties.

The CI-Test files also had to be adjusted.