Remove failing test of symbol CI
Some of the test-cases used for the CI can not be checked.
They were designed with the KLC in mind, but it was not considered if they
actually be checked by a script with reasonable effort and results.
Since its unlikely that the rules will be adjusted anytime soon to
properly handle those cases, they have been removed.