style: do not access `globals()` (relying on wildcard imports)
Previously the code for importing and using rules relied on a
combination of wildcard imports and the `globals()` dictionary.
This is a rather ugly style, which is replaced now with explicit
references and imports.

This also fixes an issue of the old (v5_schlib) implementation.
Previously warnings regarding an empty list of rules or the debug output
related to the selected rules did not use the current selection, but
rather *all* rules.