feat(kicad_sym): improve handling of invalid rotation values
Previously `NotImplemented` was raised, if the `rotation` field of a
property contained an invalid value (not one of 0/90/180/270).
Now we either raise `ValueError` (if an invalid value was given as an
argument) or we raise `NotImplementedError` (if the related instance
contains an invalid rotation value).
The latter should be improved further, when we switch to using `Enum`
classes for such fixed values.

See !429