Visual diff: exploit fast full-library rendering
This makes rendering symbols and footprints radically
faster for large change sets, as they don't have to spend
a second or two on each one to load up kicad-cli. The actual
render takes milliseconds.

It takes hundreds of items before running kicad-cli even
only two or three times is faster that rendering everything, and
it's faster than multithreading in all but very limited cases.

Once day when KiCad CLI supports multiple footprints it'll be faster
still, but with this change, it's already fast.