Add parsing of embedded files in .kicad_sym files
Since Kicad9, the user can embedd arbitrary files in the symbol. For
example a datasheet or simulation model. The parser needs to be aware
of those.

The klc-check scripts will detect those in the new rule G1.10 and report
them as error.

This commit also updates the parser to be comaptible with v9 files.