Add handling of private symbol fields
* Add KLC check for unexpected property

Print a warning when a property is encountered that is not one of the
default properties the symbol editor sets.
This commits also adds some test-symbols for the 'unit-test'

* Add handling of private symbol fields

Handling of the private property was missing in the symbol parser.
It now can serialize and de-serialize this field.
We need it for the KLC exception system, and for completeness :)

See merge request https://gitlab.com/kicad/libraries/kicad-library-utils/-/merge_requests/684