KLC checker: Add exceptions system
This allows adding a special field to (currently) symbols that lets you mark a KLC check failure as a false positive and specify a reason. That hides the failure and outputs an info message with the reason for the exception.

See merge request https://gitlab.com/kicad/libraries/kicad-library-utils/-/merge_requests/591