Fix comparelibs.py
It was broken in the commit that added the exception system.
A new parameter was added to the 'CheckSymbol' class, but in
comparelibs where an instance is created this parameter was missing.

While at it, the parameters in the caller are now named to make it
easier to understand what is going on.