Implement granular exception in S4.2
This rule is the first to support "granular exceptions". Meaning that instead
of ignoring the whole rule output - as with regular exceptions - the user can
specify what exactly will not get checked.
In this case (KLC S4.2) this is done on a pin level. By adding the private
property "KLC_S4.2_VDD{LDO}" all pins with the name "VDD{LDO}" will not get
checked by the S4.2 rule. All other pins however still can generate warnings
and errors.
In the future this can be extended to other checks as well.

This commit also adds tests for the granular and regular exception.