Fix parsing kicad_mod file with embedding 3D models
The _getArray(value) function is supposed to return an array that contains
the searched 'value' as the first item. The previous implementation did also
look for value in array items that are not the first.
That made the _getModels() function trying to parse the embedded file, which
obviously does not work.
This commit makes the _getArray() function only look at the first item. In
both the footprint and symbol parser

Resolves #422