KLC check: Relax S5.1 rule for 'generic' symbols
The rule warns if a filter is set but no default footprint exists.
This is normal for most symbols in the `Device` library. The scripts
should not complain in this case.
This change disables this warning for generic libraries.

See merge request https://gitlab.com/kicad/libraries/kicad-library-utils/-/merge_requests/701